+++ src/de/ka/javacity/graphic/IView.java
@@ -1,5 +1,5 @@
 package de.ka.javacity.graphic;
 
 public interface IView {
-	public void draw(double x, double y, int rotation);
+	public void draw(float x, float y, int rotation);
 }
