+++ src/de/ka/javacity/graphic/IView3D.java
@@ -1,5 +1,5 @@
 package de.ka.javacity.graphic;
 
 public interface IView3D {
-	public void draw(double x, double y, double z, double rx, double ry, double rz);
+	public void draw(float x, float y, float z, double rx, double ry, double rz);
 }
